http1: use size_t for the length argument of Parser::execute
`Http::Http1::Parser::execute()` declares its length argument as `int`,
but every caller passes a `size_t`:

* `ConnectionImpl::dispatchSlice()` passes `RawSlice::len_`
* the generic_proxy HTTP/1 codec passes `RawSlice::len_`
* the http_11_proxy transport socket passes `response_payload.length()`
* the http_inspector listener filter passes `new_data.length()`

Each of these narrows implicitly. A length above `INT_MAX` becomes
negative, and every guard in `BalsaParser::execute()` tests `> 0` or
`== 0`, so a negative value passes all of them and then widens back into
a very large `size_t` at `BalsaFrame::ProcessInput()`.
`LegacyHttpParserImpl` has the same shape.

This is latent rather than triggerable: read slices come from
`reserveForRead()` and never approach 2 GB, so the parser does not see
such a length today. The bound is an allocation constant rather than an
enforced invariant, so the fix is to correct the signature.

Both underlying parsers already take `size_t` -
`BalsaFrame::ProcessInput(const char*, size_t)` and
`http_parser_execute(..., size_t len)` - so widening `Parser::execute()`
to `size_t` removes the narrowing without changing behaviour, and no
call site needs to change. The `> 0` and `== 0` guards in
`BalsaParser::execute()` remain correct for an unsigned length.

Part of #46505; the `isUrlValid()` half was fixed in f08dad9cef.

+++ source/common/http/http1/legacy_parser_impl.cc
@@ -80,7 +80,7 @@ public:
     };
   }
 
-  size_t execute(const char* slice, int len) {
+  size_t execute(const char* slice, size_t len) {
     return http_parser_execute(&parser_, &settings_, slice, len);
   }
 
@@ -138,7 +138,7 @@ LegacyHttpParserImpl::LegacyHttpParserImpl(MessageType type, ParserCallbacks* da
 // same compilation unit so that the destructor has a complete definition of Impl.
 LegacyHttpParserImpl::~LegacyHttpParserImpl() = default;
 
-size_t LegacyHttpParserImpl::execute(const char* slice, int len) {
+size_t LegacyHttpParserImpl::execute(const char* slice, size_t len) {
   return impl_->execute(slice, len);
 }
